moved testing of files form the command line from TestAST.py to pyducker
